Update patch set 1
Patch Set 1:

Can you verify in drools-applications that we don't have any instance of nested classes being referred in the properties configuration?   I understand that

getName() will use "x.y.z$NestedClass" and getCanonicalName() would be "x.y.z.NestedClass".

The question is if those classes are passed through "Class.forName(..)" to instantiate them, which format does Class.forName(..) understand for nested classes, the one with "." or the one with "$" or both ?

Patch-set: 1