Update patch set 6
Patch Set 6: Code-Review+2

> Patch Set 6:
> 
> > The comment says "Say something nice", but I find that difficult
>  > right now. Is it possible that Sonar is messing with you? :-)
>  > 
>  > Seriously, it seems very undesirable to split classes into related
>  > pairs like this. The grouping of API methods and associated
>  > constants seems very natural, and this separation seems very
>  > kludgy.
> 
> I kind of agree, though on the other hand, I don't like to see interfaces cluttered with things other than methods.  In any event, this seems to be more of a preference thing, so decided to just do it the way sonar demanded.

Easier to just satisfy the sonar versus personal preferences.
According to Joshua Bloch, author of "Effective Java":

The constant interface pattern is a poor use of interfaces.

That a class uses some constants internally is an implementation detail.

Implementing a constant interface causes this implementation detail to leak into the class's exported API. It is of no consequence to the users of a class that the class implements a constant interface. In fact, it may even confuse them. Worse, it represents a commitment: if in a future release the class is modified so that it no longer needs to use the constants, it still must implement the interface to ensure binary compatibility. If a nonfinal class implements a constant interface,

all of its subclasses will have their namespaces polluted by the constants in the interface.

Patch-set: 6
Label: Code-Review=+2