Update patch set 3
Patch Set 3:

> (1 comment)

It is my impression, and let me know what you think, because stubbing out in the fsm the sheduledthreadpooolexecutor with a test version, there is effectively a loss of ability to find problems simply because the executor implemention underneath is just different, even though the coverage % stays the same.   I don't remember right now all the tests, but my concern is that the timer functionality cancel, starting, etc .. when doing state changes may not be tested properly, which makes me a little nervous.  As a note it takes little more than one minute in my vm to run all junits in this module, I think the delays would be in the active-standby module.    Anyhow, let me know what you think, but that's where my concern is coming from.

Patch-set: 3