Update patch set 3
Patch Set 3:

> > (1 comment)
 > 
 > It is my impression, and let me know what you think, because
 > stubbing out in the fsm the sheduledthreadpooolexecutor with a test
 > version, there is effectively a loss of ability to find problems
 > simply because the executor implemention underneath is just
 > different, even though the coverage % stays the same.   I don't
 > remember right now all the tests, but my concern is that the timer
 > functionality cancel, starting, etc .. when doing state changes may
 > not be tested properly, which makes me a little nervous.  As a note
 > it takes little more than one minute in my vm to run all junits in
 > this module, I think the delays would be in the active-standby
 > module.    Anyhow, let me know what you think, but that's where my
 > concern is coming from.

I guess I will leave that up to you.  I'm pretty comfortable using simulators, though it does leave a small window through which errors may slip.  In the case of feature-lifecycle, that doesn't seem too likely, but who knows.
As for the test only taking a minute, I like to be able to make changes and quickly run a test to verify that the changes didn't break anything.  If I have to wait a minute for the test to run, then I probably won't bother running it at all, but will just let jenkins do it; long tests are just not very CI/CD-friendly IMHO.  I tend to be rather impatient - "long" to me is more than 5-10 seconds.
Nevertheless, if you prefer, I will revert the changes to feature-lifecycle.

Patch-set: 3