Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Can you mark the JIRA in progress - as well as ensure it really should be standalone vs a sub-task within a user-story and/or epic. Thanks.

I have marked it "in-progress".

This change was actually prompted by a bug in an ECOMP control loop -- this capability is needed to fix the problem. The code change here makes it a generic operation, so it isn't just something in server pool. It is similar to the new lock API in that respect.

Patch-set: 1