Update patch set 1
Patch Set 1:

> > Patch Set 1:
 > >
 > > I think PolicyContainer.insert(..) and PolicyContainer.insertAll(..)
 > should be updated to call PolicySession.insertDrools(..) to make
 > sure the interception mechanism happens for all insertion paths.
 > 
 > Do we want to support both direct and smart inserts, or should only
 > the smart insert be provided. Would this affect incoming UEB/DMAAP
 > inserts as well? In server pool, I am currently intercepting
 > 'PolicyControllerFeatureApi.beforeOffer' in order to forward
 > messages -- if incoming UEB/DMAAP messages did smart inserts,
 > things would still work without the 'beforeOffer' intercept.

The changes I am suggesting will provide hooks to all inserts, so both direct and smart, so all invokations to kiesession.insert(..) will have the possibility to be intercepted regardless.   The "smart" vs. "non-smart" insert decision will be up to the underlying feature to make that distinction.    So, this will affect all paths including the UEB/DMaaP and inserts through telemetry APIs.   It seems that the beforeOffer won't be necessary in server pool may be redundant then, unless you want to do something special directly withthe PolicyController object that is passed in (although could be inferred too in a slightly more complex way from PolicySession).    I think the main benefit for the change would be consistency, and maintenance, as having all kie inserts in the same place in those 2 methods in the PolicySession, instead of in several places, ie. PolicyContainer and PolicySession.

Patch-set: 1
Label: Verified=0