Update patch set 20
Patch Set 20: -Code-Review

(5 comments)

If there is a specific manager class for each feature implementation, then I don't see the value in defining a feature API at all, as manager class would do the work.
Note: whatever approach is taken, the implementation must still be called when a controller is started so that it can configure itself with controller-specific properties.

Patch-set: 20