Update patch set 27
Patch Set 27:

> (1 comment)

The concern I have is with leaving resources hanging in the db preventing an operation when a pdp-d goes away.   The stop/shutdown operation is a graceful operation perform by an admin, where the liveness check for example would kill the container, in the latter saving existing transactions is relevant.

In ONAP, persistence feature is not enabled and would be wary to activate it due to its poor performance and other unknowns. Ralph's server pooling has its own more efficient persistence though.   For ONAP though there is this persistence epic thing to work across PDPs, so I think it is safe to  that assume persistence is just not there at this point, I would think this functionality will take quite a while to get  straight and mature.   In the meantime, I advocate to not leave in progress transactions.   Actually, the works that dcae works for us, as the alarms on the same resource are retransmitted in few millisecs batch, it would work best for the PDP going down to try to gracefully clean up before going away, so the next pdp available picks up that onset right away, and starts reprocessing it.

In any case, this is clearly out of scope for this work for sure, but good context to discuss.

Patch-set: 27