Update patch set 5
Patch Set 5:

> (1 comment)

I don't think this AES_ENCRYPTION_KEY would be used in ONAP, neither in policy/common nor the drools repos, and most likely would be extended to other repos as well.   

The rationale is that the ONAP policy/common as it should be kept as *generic* libraries used across different components should not impose *installation requirements in the OS* via specific naming and the use of environment variables and forcing it to be preset by the installer in the OS.   Another component depending on its characteristics may choose to get the key in some other way and not through an env. variable.

Another argument is that we want to keep the implementation somewhat open and open it to other folks to other crypto mechanisms, for example if somebody wants to use DES, the name would serve no purpose and be confusing.

Note that the benefits of this review are huge compared with the original proposal, as it has been made transparent to the all drools and drools-applications code.   Any sensitive configuration data regardless would be handle transparently to the application.   No more individual calls to cryptoutils all over the place on a per individual field basis that require code and configuration changes.   The symm key just need to base in the base.conf and the encoded data in the *.conf where you wanted with no code changes.   Again the maintenance benefits are of significance in this implementation compared with the originally proposed.

Patch-set: 5