Update patch set 3
Patch Set 3:

> > > (1 comment)
 > >
 > > I think if we remove the final to make it easy in junits, I think
 > > we could probably go all the way, so we are weakening the
 > > constraint, but still imposing not being null at any given time.
 > 
 > That's fine.  However, we cannot impose "NotNull", because it may
 > be null, depending on which junit test class is run and in what
 > order.  In fact, I think it's null if this particular test is run
 > by itself.

couldn't junits simply set a do nothing "NullPolicyEngine" or any implementation anybody wants, o via reflection set it explicitly to null (I think what's the Whitebox statement from Ralph probably uses reflection).

In production code, we always want to maintain the invariant that there is always a non-null PolicyEgine implemention.   If it's open with a setter for generic junit support, the @NonNull annotation in the method should enforce that.   Provided that anybody can use reflection to modify it, but is not part of the contract.

Patch-set: 3