upgrade to 7.28.0.Final drools libraries
commons-lang does not seem to be packaged anymore
(commons-lang3 still does, which is the desired one).

There are some @SupressWarning that seem to have no effect
that were removed.

Some libraries that were explicitly added are transitively
used now in the latest version of drools, hence some rework
in the dependency and exclude sections.   We will need to
generate a clm report to verify.

Issue-ID: POLICY-1407
Signed-off-by: jhh <jorge.hernandez-herrero@att.com>
Change-Id: I44e09183caca72f5f33bdec0b8fa88b0b4793055
Signed-off-by: jhh <jorge.hernandez-herrero@att.com>