Update patch set 1
Patch Set 1: Code-Review-1

Sorry Jim, I think we have to go back to the drawing table on this, these changes imply re-architecture changes over previous releases that I need to understand better as they are system disrupting, as the whole notion of advertising capabilities does fully go away, and thought it would be preserved.   I don't know how it will combine too with the native policies, where they could just add any arbitrary "rules" artifacts.

This code served a very concrete purpose in last releases, as the propagation of the "capabilities" ie. policy types based on this happened only when the kie resources where fully ready, so in essence delaying getting assignments of policies until then.   Note that initialization of underlying of drools resources may take quite a bit of time, and especially when the native policies are introduced.   There was also a mapping to the appropriate controller base on policy types.   I need to get a better grasp on it.

As an aside, it would be good to document all these changes in the Policy Framework architecture documentation so it is kept up to date.

I suggest marking this review as WIP and postpone until we all come back from vacations.   It would be nice to see a write up with all architecture changes that are coming in in this release that are stemming from the multiple group support, there are quite a bit.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1