Update patch set 3
Patch Set 3:

May be I am not fully aware of drools-pdp code base, but I didn't find a class where drools is creating/sending the PdpStatus message to PAP. 
Hence, not clear if supported policy types is ONLY removed from the PdpStatus and will still be maintained in the drools-pdp. So that the next Jira on "Resync of PAP-PDP when supported policy types of a PdpSubGroup changes" can be implemented.

Patch-set: 3