Update patch set 4
Patch Set 4:

Thank you folks.  In summary, what we are saying is that overall validation strategy across components would be revisited in G. release and Pam open POLICY-2352 ticket for it.   I'll go ahead and merge this one, and cover a few comments from Jim and Chenfei in a different review.

Patch-set: 4
Label: Code-Review=0