Native policies support in docker
Depends on https://gerrit.onap.org/r/#/c/policy/docker/+/101758/
for package cleanup.

The work is needed to run the PDP-D with no applications, native
policies may exercise this set up.   This work provides a functional
image with a precooked installation.

This work will allow fixing and cleaning up of CSITs specific to
this repo which maintenance is problematic.

Issue-ID: POLICY-2181
Signed-off-by: jhh <jorge.hernandez-herrero@att.com>
Change-Id: I62a1b9b556bd2f6c0c243062cb864dfed80e2bea