Update patch set 3
Patch Set 3:

> Not sure why we need this class Jim, provided that there is an
 > executor in the PolicyEngine.  In general, I am not sure I like the
 > proliferation of executor in any place, and then dealing with the
 > canceling of the tasks, etc .. in shutdown places from multiple
 > places

The purpose of this class is to execute tasks within the drools engine (i.e., KieSession, not PolicyEngine).  Drools-pdp won't use it, but drools-apps will create instances and pass them to the Actors as an "Executor" so that when the Actors invoke callbacks, the callbacks will be executed within the drools engine.  The executor contained within PolicyEngine is for asynchronous tasks and does not execute them within the drools engine.  I could move this class to drools-apps, but it kind of makes sense as a general-purpose utility since it's closely tied to DroolsRunnable, which is also in drools-pdp/core.

Patch-set: 3