Update patch set 3
Patch Set 3: Code-Review+1

> > Not sure why we need this class Jim, provided that there is an
 > > executor in the PolicyEngine.  In general, I am not sure I like
 > the
 > > proliferation of executor in any place, and then dealing with the
 > > canceling of the tasks, etc .. in shutdown places from multiple
 > > places
 > 
 > The purpose of this class is to execute tasks within the drools
 > engine (i.e., KieSession, not PolicyEngine).  Drools-pdp won't use
 > it, but drools-apps will create instances and pass them to the
 > Actors as an "Executor" so that when the Actors invoke callbacks,
 > the callbacks will be executed within the drools engine.  The
 > executor contained within PolicyEngine is for asynchronous tasks
 > and does not execute them within the drools engine.  I could move
 > this class to drools-apps, but it kind of makes sense as a
 > general-purpose utility since it's closely tied to DroolsRunnable,
 > which is also in drools-pdp/core.

Thanks for the feedback, I think it is ok here then.   It always bugged me (this is previous to your work) that the actors were doing manipulation of the working memory, but may be there is no way to work around it (?).   One issue if the actors depend on drools WorkingMemory that I wonder, if it imposes restrictions to be used outside of drools for example in apex.

Patch-set: 3
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1