Update patch set 18
Patch Set 18:

Ram/Liam - just a reminder from the discuss months ago. It was agreed this code was to be submitted and then further testing would determine which pooling strategy Drools would use overall. Work to be finished in the G Release. Bobby is creating JIRA's for finishing all the work.

Patch-set: 18