Update patch set 1
Patch Set 1:

I am not sure if this will solve the problem that it causes the failure in the merge, I see the merge failure complaining a couple lines below.

java.lang.AssertionError
	at org.onap.policy.drools.serverpooltest.Test1.sessionMigrationTest(Test1.java:445)

I occasionally see some merge errors in the m2 module tests in the apps side, we may need you to take a look at it too.

The test seem to take between 1-1.5 mins, regardless sucess or failure, so the 5 min. seem excessive:

[INFO] feature-server-pool ................................ SUCCESS [01:36 min]
---
[INFO] feature-server-pool ................................ FAILURE [01:47 min]

Patch-set: 1
Label: Verified=0