Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > (1 comment)
> 
> I will make all server IP for junit configurable from property file.

Joseph, keep in mind that we don't want developers to have to configure their environment just to build the repo. 'mvn clean install' should work out-of-the-box as long as their environment is reasonable (contains JDK 11, fairly new OS version).

Is there anyway we just use 127.0.0.1 for all these cases but dynamically alloc ports for binding?

You can consider using a TemporaryFolder to copy a properties file if you know that you would need to change it for JUnits.

Patch-set: 3
Label: Code-Review=0