Update patch set 1
Patch Set 1: Code-Review-1

> This just isn't good timing as we are about to release images. I'd
 > like to wait until after we release the current version to
 > Integration.
 > 
 > Jorge please let me know your thoughts

Yes, I agree, this would have to be postponed until we release our first set of images.

I would also ask to include more documentation on what the code changes are correcting, on the JIRA ticket, so we don't need to verify ourselves.   Full compilation of the other repos should be performed as well to minimize the possibility of breakage.  This is to avoid the problems that the previous log4j (rejected) submission had.  Transitive dependencies problems are the most difficult problems to debug if they cause a problem.

I also think the code change is incorrect, running a mvn dependency:tree in policy-core, kie-api does not have the ant dependency (kie-ci instead does, so the code change should go there).   Let's document that in the ticket next time Joseph, so we, committers, don't need to do it ourselves.

Lastly, let's have the summary line (first line) descriptive of the commit change. For example "ONAP" word is unnecessary, this repo is ONAP.   Please refer to https://wiki.onap.org/display/DW/Commit+Messages for guidelines.   It seems that the second line is more representative of the work, and could replace the first one in this commit for example (under 50 chars per guidelines in wiki page).

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1