Update patch set 16
Patch Set 16:

> Patch Set 15: -Code-Review
> 
> > Patch Set 15:
> > 
> > (1 comment)
> 
> Isn't it better to just leave the logback files out altogether?  Can include them in the CSITs and oom instead, right, where they need not include all of these extra fields?  Companies can add their own for deployment via kubernetes, right?

My comments are along making sure the default logback.xml is consistent with the base code in the image.

I think in the OOM if the defaults are OK, I don't see a major reason to create a duplicate of it to just put it outside.   I think it is documented how to override it (I'll make sure it is if not), which is pretty straighforward.   For CSITs, I can see the argument more clear to put it in the mounted directory as it is more used for development, and don't have the size limitations of the charts.

Patch-set: 16
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>