Update patch set 2
Patch Set 2:

(7 comments)

This is good, but have some concerns with usability by other new types of controllers besides tdjam, for example, the DMN controllers that would be introduced in this release.   I suggest to mark this review as work in progress as it is a very good base, and I can work with Taka and folks to revisit a few things to to incorporate other controllers besides tdjam.

Patch-set: 2
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>