Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > should this change be done here, or in policy/docker?
> > 
> > It would be better in policy/docker so all inherit it ..
> 
> which portion would you suggest to move to policy/docker? jdk docker and telemetry tool is only using drools-pdp now...

I am guessing that "jq" that is installed in the policy-jdk has a dependency with httpie, that's the reason I am guessing you need to do a pip"uninstall" (?) of it followed by an "install" of it to delete the other version that "jq" installed as a dependency through the regular alpine package manager in the base image.   That's the reason for the suggestion to change it in the base but I am fine too, as Jim brings up in another comment to leave the package override just in drools.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>