sort deploy and undeploy policy lists
An order is imposed in the deployment and undeployment actions when
the list of active policies is retrieved from PAP.   This is to
ensure that the operations are applied in a sane way, for example
to try to prevent to undeploy policies before deleting a controller,
etc ..

The deployment order is 1) native controller, 2) native rule, and 3)
non-native policies.

The undeployment order is 1) non-native, 2) native rule, and 3)
native controller policies.

Issue-ID: POLICY-2762
Signed-off-by: jhh <jorge.hernandez-herrero@att.com>
Change-Id: Icb15dcec87fd5d9917ee152ab15ca7277e13590a