sync policies when native artifact policies added
interoperability between native and non-native policies
present several difficulties.   The code submitted
operates in deltas where deploy and undeploy operations
are performed on the deltas of the updates.

In order to support interoperability, policies not
part of a delta update may need to be reapplied.
For example, in the case when the delta is just
a new native controller artifact, we should go
through the set of already deployed policies and
reapply non-native policies as long as the policy types that
the native artifact policy supports.

Issue-ID: POLICY-2762
Signed-off-by: jhh <jorge.hernandez-herrero@att.com>
Change-Id: Ieb8e5e17862e9d607433a1d4e86a026725d73498
Signed-off-by: jhh <jorge.hernandez-herrero@att.com>