Fix sonars from dependency upgrade
The dependency upgrades in policy-parent caused some new sonars.  In
particular, initMocks() has been deprecated - replaced with calls to the
Mockito Runner.  Changed one to openMocks() because using the Runner
caused it to fail on jenkins, but not in the local eclipse.
Also fixed a type safety issue with EMPTY_LIST.

Issue-ID: POLICY-2909
Change-Id: Iea9420190f208d7de3bf3ad531515c92e0fab83f
Signed-off-by: Jim Hahn <jrh3@att.com>