Upgrade hibernate dependency version in drools-pdp
The FOSS report for drools-pdp showed that hibernate-commons-annotations
uses the LGPL 2 license, which means it should be kept up to date, if
possible.  However, it is still version 5.0.1.Final, which was released
in 2015, while 5.1.2.Final was released in 2020.  Upgraded to the new
version.
Verified junits still pass in both drools-pdp & drools-apps.

Issue-ID: POLICY-3005
Change-Id: I1c42b14cfa812299797de002a7fa2e20b49dea6e
Signed-off-by: Jim Hahn <jrh3@att.com>