Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> (1 comment)
> 
> while reviewing the changes to apex-pdp, we made the decision that, when responding to PAP, the PDPs should only send back the original deployment list (with failures removed), instead of sending back the complete list of policies.  I don't think I saw that change here.  Maybe that will be done in a separate review?

I didn't realize it in the apex review,  I have some follow up questions/concerns: 

1.
We are lacking documentation for this feature, and in general the process to grow a PDP type from PAP perspective, what is expected of a new type of PDP to comply with to be recognized by PAP.

2.
What needs to be sent back exactly in the response, the comment refers to the deploy list but there are 2 lists, deploy, and undeploy, that could experiment failures.

3.
What's the reason PAP requires this changes wrt current behavior, how does PAP handles the response and acts upon failures reported on it?

4.
What are the rules for STATUS messages triggered from heartbeat, any changes related to the policy list being sent out?   It may have been good to have an specific RESPONSE message to avoid overloading the STATUS message processing conditional logic to act differently depending in context.

5.
In "legacy" processing, PAP could be sync'ed with PDP status at all times as the STATUS messages contain the list of policy ids.    How consistency is achieved by the PAP with the delta policy?   Does the UPDATE "policies" field need to be used for any purpose, perhaps re-synchronization, etc .. or operator command to resync a PDP with PAP?

Patch-set: 2