Update patch set 1
Patch Set 1:

In principle these classes are direct mappings of the policy-models *common.Drools policy type, so should only change when the policy type in policy-models change to include a guard disable property.   They were initially generated from the jsonchema under src/main/resources and tweaked a little.  The jsonschema is what is coupled with the actual policy type definition and is what it provides policy validation against its policy type and domain conversion.   So in theory, any changes here, should be accompanied by changes in policy-models related policy type and its counterpart jsonschema for drools (for the validation piece).

This would also apply on per policy basis, so this will require all tosca policies to carry the disable flag in the properties to be honored, more of a reason require the changes in policy-models.

In the past we had the "guard.disabled" flag as an environment variable to provide this functionality but system wide, it was a before the actor redesign, but I don't remember what happen, I think we talked about it, but I don't remember the outcome, if it was discontinued or not.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>