Update patch set 3
Patch Set 3:

I think it is only drools, the other components seem to have the port configurable or expose the jdbc url externally for configuration.   I am not sure about apex, but they may not be a need there, question for apex folks .. I see references to postgres database running on localhost.

Patch-set: 3