Remove GPLv3 components from drools-pdp
The drools-pdp docker image includes GNU tar, which is GPLv3 and is
therefore disallowed by ONAP.  Appears to be included with "dpkg", which
isn't needed anymore, so just removed it.
Also removed the code that installs python, as it's now included in the
policy jdk image.
Also changed policy/docker version, as the prior version did not include
python3.

Issue-ID: POLICY-2845
Change-Id: I08b97b90aef6820135b052c20f270f8b487c94a9
Signed-off-by: Jim Hahn <jrh3@att.com>