Reduce coder verbosity in logs
After switching to lombok ToString, the log messages for the coder class
have become very long.  Switched back to the original toString() method,
which only logs the keysets rather than the whole content of the coder
maps.
Left the other additions of lombok ToString in drools-pdp as is, because
they all looked appropriate, upon further review.

Issue-ID: POLICY-3397
Change-Id: I08ea8c49908b33fc3b8c8ba6de4ac4bd1474660a
Signed-off-by: Jim Hahn <jrh3@att.com>