Update patch set 3
Patch Set 3:

(3 comments)

In addition to this, we need to perform CSITs to make sure that the policy handling from API<---->PAP<----PDPs continues to work correctly.

Also, although this commit is building in gerrit, it is not building locally for either Ramesh or me, The test RestLifecycleManagerTest.testMultiPolicyFlow() when run on its own succeeds, when we run all the tests in feature-lifecycle, this test fails. It seems that somehow, the fsm.shutDown() is getting called before the test can finish.

Patch-set: 3
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>