Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Jorge - could this be added to the other PDPs?

It can be added at least the script to dump them, the other functionality in drools can be a bit of work, but somewhat equivalent functionality could be added.   The way that it works for drools, the periodic kubernetes liveness checks will trigger healthy checks in the drools controllers, this is by checking on the underlying kiesessions (we have seen these problems with unresponsive drools applications before).  If detected that a kiesession for that controller is stuck, the drools pod will automatically be restarted after 3 liveness probes failures to get unblocked from that situation where the drools application is unresponsive.   At the same time when the controller "stuck" condition is detected, per this review, it will issue a thread dump, which can then be analyzed postmortem as part of the logs on how we got into this situation.     I can add a task to the xacml pdp epic, and we can decide if we want to do something similar in other components.

Patch-set: 2