Update patch set 2
Patch Set 2:

> question, is there an url that can generate the swagger spec?


At the moment there is no endpoint. In the other spring projects we used spring-doc that creates /v3/api-docs endpoint.

In this case, during the build a swagger.json file is generated in target/generated-sources/swagger/swagger.json I could add an endpoint GET /swagger to retrieve this document, and change the endpoint in the telemetry tool.

Patch-set: 2